Refactor alert item into its own component.
