+++ src/components/map/StopPopupContent.js
@@ -3,7 +3,7 @@ import StopRouteSelect from "./StopRouteSelect";
 import {Text} from "../../helpers/text";
 import React from "react";
 import styled from "styled-components";
-import AlertsList, {AlertItem} from "../AlertsList";
+import AlertsList from "../AlertsList";
 import {StopContent} from "../StopElements";
 import {getAlertsInEffect} from "../../helpers/getAlertsInEffect";
 import {Button} from "../Forms";
@@ -11,10 +11,6 @@ import {Button} from "../Forms";
 export const StopAlerts = styled(AlertsList)`
   padding: 0;
   margin-top: 1rem;
-
-  ${AlertItem} {
-    font-family: inherit;
-  }
 `;
 
 export const StopPopupContentSection = styled.div`
